Reverse Coding Guide
This is a reverse coding guide for every possible HiTOP item in the MIDUS data set. The aim of this guide is to ensure that the highest value in each scale represents the largest quantity of the trait. This means that, for instance, pain and pleasure will correlate because the scale is coded such that higher scores on each scale represent more of the trait. To change this, further coding is necessary. For reverse coding specific to the scales developed, please see the finalized code and scale guidelines. This guide is simply a data preparation guide before imputation and initial analysis.