Parallel EFA
This is the R script for an EFA based on parallel analysis of the whole item pool, without iteration. Visual analysis of the scree plot suggested a 5 factor solution could fit well, despite officially recommending over 20 factors. This script will identify, based on the EFA with 5 factors, an item pool for the CFA with items assigned to factors based on their highest loading. Each item must have a loading above |.30|, and a communality above .25 to be assigned to a factor. The rest of the items are dropped and logged.